requirements.txt
asgiref==3.4.1
astroid==2.6.2
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-environ==0.4.5
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
drfdocs==0.0.11
idna==2.10
inflection==0.5.1
isort==5.9.2
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==21.0
pylint==2.9.3
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.2
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
psycopg2-binary
psycopg2
drf-yasg